Menambahkan sistem tema dan penambahan banyak fitur baru

+++ app/Http/Controllers/ProfileController.php
@@ -2,6 +2,8 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\User;
+use Hash;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Storage;
@@ -27,9 +29,11 @@ class ProfileController extends Controller
         ]);
         $user = Auth::user();
         if ($req->hasFile('avatar')) {
-            $pathFileToRemove = Str::after($user->avatar, '/storage/');
-            if (Storage::disk('public')->exists($pathFileToRemove)) {
-                Storage::disk('public')->delete($pathFileToRemove);
+            if ($user->avatar) {
+                $pathFileToRemove = Str::after($user->avatar, '/storage/');
+                if (Storage::disk('public')->exists($pathFileToRemove)) {
+                    Storage::disk('public')->delete($pathFileToRemove);
+                }
             }
             $path = $req->file('avatar')->store('avatars', 'public');
             $user->avatar = env('APP_URL') . "/storage/$path";
@@ -39,5 +43,65 @@ class ProfileController extends Controller
         $user->organization = $validated['organization'];
         $user->save();
         return redirect()->route('profile')->with('success', 'Profile updated successfully.');
+    }
+    public function change_password()
+    {
+        $user = Auth::user();
+        if ($user->provider == "google") {
+            return redirect('/profile')->with('error', "Can't change the password because you're account from google.");
+        }
+        return Inertia::render('profile/change-password');
+    }
+    public function change_password_process(Request $req)
+    {
+        $validated = $req->validate([
+            'newPassword' => 'required|string|confirmed',
+            'currentPassword' => 'required|string|current_password',
+        ]);
+        $user = Auth::user();
+        User::query()->where('id', $user->id)->update([
+            'password' => Hash::make($validated['newPassword']),
+        ]);
+        return redirect('/profile')->with('success', 'Success to change the password.');
+    }
+    public function delete_account(Request $req)
+    {
+        $req->validate([
+            'password' => 'required|current_password'
+        ]);
+        $user = $req->user();
+        if ($user->avatar) {
+            $pathFileToRemove = Str::after($user->avatar, '/storage/');
+            if (Storage::disk('public')->exists($pathFileToRemove)) {
+                Storage::disk('public')->delete($pathFileToRemove);
+            }
+        }
+        Auth::logout();
+        $user->delete();
+        $req->session()->invalidate();
+        $req->session()->regenerateToken();
+        return redirect('/login');
+    }
+    public function save_settings(Request $req)
+    {
+        $validated = $req->validate([
+            'theme' => 'required',
+            'accent' => 'required',
+            'density' => 'required',
+            'auto_save' => 'required|boolean',
+        ]);
+        $data = [
+            'auto_save' => $validated['auto_save'],
+            'appearance' => [
+                'theme' => $validated['theme'],
+                'accent' => $validated['accent'],
+                'density' => $validated['density'],
+            ],
+            'notifications' => true,
+        ];
+        $user = Auth::user();
+        $user->settings = json_encode($data);
+        $user->save();
+        return back();
     }
 }
